third_party/WebKit/LayoutTests/inspector/elements/elements-panel-search.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html id="documentElement">
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>
function initializeShadowDOM()
{
    var shadow = document.querySelector('#shadow-host').createShadowRoot();
    var template = document.querySelector('#shadow-dom-template');

    // Avoid matching this function
    shadow.appendChild(template.content.cloneNode(true));
}

function test()
{
    var omitInnerHTML;

    function searchCallback(next, resultCount)
    {
        if (resultCount == 0) {
            InspectorTest.addResult("Nothing found");
            SDK.DOMModel.cancelSearch();
            next();
        }

        for (var i = 0; i < resultCount; ++i)
            InspectorTest.domModel.searchResult(i, searchResultCallback.bind(this, i + 1 === resultCount));

        function searchResultCallback(isLastItem, node)
        {
            node.getOuterHTML(addSearchResult.bind(this, isLastItem));
        }

        function addSearchResult(isLastItem, error, markupVa_lue)
        {
            if (omitInnerHTML)
                markupVa_lue = markupVa_lue.substr(0, markupVa_lue.indexOf(">") + 1);
            InspectorTest.addResult(markupVa_lue.split("").join(" "));
            if (isLastItem) {
                SDK.DOMModel.cancelSearch();
                next();
            }
        }
    }

    function setUp(next)
    {
        InspectorTest.domModel.requestDocument(step2);

        function step2()
        {
            InspectorTest.evaluateInPage("initializeShadowDOM()", next);
        }
    }

    InspectorTest.runTestSuite([
        function testSetUp(next)
        {
            setUp(next);
        },

        function testPlainText(next)
        {
            InspectorTest.domModel.performSearch("Fo" + "o" + "Bar", false, searchCallback.bind(this, next));
        },

        function testPartialText(next)
        {
            InspectorTest.domModel.performSearch("oo" + "Ba", false, searchCallback.bind(this, next));
        },

        function testStartTag(next)
        {
            InspectorTest.domModel.performSearch("<inpu" + "t", false, searchCallback.bind(this, next));
        },

        function testEndTag(next)
        {
            InspectorTest.domModel.performSearch("npu" + "t>", false, searchCallback.bind(this, next));
        },

        function testPartialTag(next)
        {
            InspectorTest.domModel.performSearch("npu" + "t", false, searchCallback.bind(this, next));
        },

        function testPartialAbsentTagStart(next)
        {
            InspectorTest.domModel.performSearch("<npu" + "t", false, searchCallback.bind(this, next));
        },

        function testPartialAbsentTagEnd(next)
        {
            InspectorTest.domModel.performSearch("npu" + ">", false, searchCallback.bind(this, next));
        },

        function testFullTag(next)
        {
            InspectorTest.domModel.performSearch("<inpu" + "t>", false, searchCallback.bind(this, next));
        },

        function testExactAttributeName(next)
        {
            InspectorTest.domModel.performSearch("valu" + "e", false, searchCallback.bind(this, next));
        },

        function testExactAttributeVal_ue(next)
        {
            InspectorTest.domModel.performSearch("In" + "putVa" + "l", false, searchCallback.bind(this, next));
        },

        function testExactAttributeVal_ueOnRoot(next)
        {
            omitInnerHTML = true;
            InspectorTest.domModel.performSearch("documen" + "tElement", false, searchCallback.bind(this, next));
        },

        function testExactAttributeVal_ueWithQuotes(next)
        {
            omitInnerHTML = false;
            InspectorTest.domModel.performSearch("\"fo" + "o\"", false, searchCallback.bind(this, next));
        },

        function testPartialAttributeVal_ue(next)
        {
            InspectorTest.domModel.performSearch("n" + "putVa" + "l", false, searchCallback.bind(this, next));
        },

        function testXPathAttribute(next)
        {
            InspectorTest.domModel.performSearch("//html" + "//@attr", false, searchCallback.bind(this, next));
        },

        function testSelector(next)
        {
            InspectorTest.domModel.performSearch("d" + "iv.divclass span", false, searchCallback.bind(this, next));
        },

        function testCaseUpperFindsLower(next)
        {
            InspectorTest.domModel.performSearch("K" +"OO", false, searchCallback.bind(this, next));
        },

        function testCaseLowerFindsUpper(next)
        {
            InspectorTest.domModel.performSearch("c" + "aseless", false, searchCallback.bind(this, next));
        },

        function testCaseAttribute(next)
        {
            InspectorTest.domModel.performSearch("C" + "AMEL", false, searchCallback.bind(this, next));
        },

        function testSearchShadowDOM(next)
        {
            InspectorTest.domModel.performSearch("<c" + "ontent", false, searchCallback.bind(this, next));
        },

        function testSearchUAShadowDOM(next)
        {
            InspectorTest.addResult("Searching UA shadow DOM with setting disabled:")
            InspectorTest.domModel.performSearch("inne" + "r-editor", false, searchCallback.bind(this, step2));

            function step2()
            {
                InspectorTest.addResult("Searching UA shadow DOM with setting enabled:")
                InspectorTest.domModel.performSearch("inne" + "r-editor", true, searchCallback.bind(this, next));
            }
        },

        function testSearchShadowHostChildren(next)
        {
            InspectorTest.domModel.performSearch("shadow-host-c" + "ontent", false, searchCallback.bind(this, next));
        },
    ]);

}
</script>
</head>

<body onload="runTest()">
<p>
Tests that elements panel search is returning proper results.
</p>

<div>FooBar</div>
<input value="InputVal">
<div attr="foo"></div>
<div id="terminator"></div>
<div class="divclass"><span>Found by selector</span></div>
<span class="foo koo"></span>
<span class="CASELESS"></span>
<span data-camel="insenstive"></span>
<div id="shadow-host">
    <div id="shadow-host-content"></div>
</div>
<template id="shadow-dom-template">
  <div id="shadow-dom-outer">
      <content></content>
  </div>
</div>
</template>
<textarea></textarea>

</body>
</html>
